fix: move listeners, jobs, and service into the application layer.
